Reply by email subject values
I’d like to submit this change which still assigns the post title to the $subject, otherwise it will assign the permalink instead.  This would be more useful for say an all-media or quote post that doesn’t start with traditional plain text, and would make it easier to reference back to the actual link.